Use u64 for integers in algonaut_algod

+++ algonaut_algod/src/models/account_participation.rs
@@ -10,8 +10,6 @@
 
 /// AccountParticipation : AccountParticipation describes the parameters used by this account in consensus protocol.
 
-
-
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct AccountParticipation {
     /// \\[sel\\] Selection public key (if any) currently registered for this round.
@@ -22,13 +20,13 @@ pub struct AccountParticipation {
     pub state_proof_key: Option<String>,
     /// \\[voteFst\\] First round for which this participation is valid.
     #[serde(rename = "vote-first-valid")]
-    pub vote_first_valid: i32,
+    pub vote_first_valid: u64,
     /// \\[voteKD\\] Number of subkeys in each batch of participation keys.
     #[serde(rename = "vote-key-dilution")]
-    pub vote_key_dilution: i32,
+    pub vote_key_dilution: u64,
     /// \\[voteLst\\] Last round for which this participation is valid.
     #[serde(rename = "vote-last-valid")]
-    pub vote_last_valid: i32,
+    pub vote_last_valid: u64,
     /// \\[vote\\] root participation public key (if any) currently registered for this round.
     #[serde(rename = "vote-participation-key")]
     pub vote_participation_key: String,
@@ -36,7 +34,13 @@ pub struct AccountParticipation {
 
 impl AccountParticipation {
     /// AccountParticipation describes the parameters used by this account in consensus protocol.
-    pub fn new(selection_participation_key: String, vote_first_valid: i32, vote_key_dilution: i32, vote_last_valid: i32, vote_participation_key: String) -> AccountParticipation {
+    pub fn new(
+        selection_participation_key: String,
+        vote_first_valid: u64,
+        vote_key_dilution: u64,
+        vote_last_valid: u64,
+        vote_participation_key: String,
+    ) -> AccountParticipation {
         AccountParticipation {
             selection_participation_key,
             state_proof_key: None,
@@ -47,5 +51,3 @@ impl AccountParticipation {
         }
     }
 }
-
-
