+++ algonaut_algod/src/models/dryrun_source.rs
@@ -10,24 +10,22 @@
 
 /// DryrunSource : DryrunSource is TEAL source text that gets uploaded, compiled, and inserted into transactions or application state.
 
-
-
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct DryrunSource {
     #[serde(rename = "app-index")]
-    pub app_index: i32,
+    pub app_index: u64,
     /// FieldName is what kind of sources this is. If lsig then it goes into the transactions[this.TxnIndex].LogicSig. If approv or clearp it goes into the Approval Program or Clear State Program of application[this.AppIndex].
     #[serde(rename = "field-name")]
     pub field_name: String,
     #[serde(rename = "source")]
     pub source: String,
     #[serde(rename = "txn-index")]
-    pub txn_index: i32,
+    pub txn_index: u64,
 }
 
 impl DryrunSource {
     /// DryrunSource is TEAL source text that gets uploaded, compiled, and inserted into transactions or application state.
-    pub fn new(app_index: i32, field_name: String, source: String, txn_index: i32) -> DryrunSource {
+    pub fn new(app_index: u64, field_name: String, source: String, txn_index: u64) -> DryrunSource {
         DryrunSource {
             app_index,
             field_name,
@@ -36,5 +34,3 @@ impl DryrunSource {
         }
     }
 }
-
-
